doc: update readme.md

+++ data/stylegan3/dataset.py
@@ -115,7 +115,7 @@ class Dataset(torch.utils.data.Dataset):
         if self._xflip[idx]:
             assert image.ndim == 3 # CHW
             image = image[:, :, ::-1]
-        return image.copy() / 255., self.get_label(idx)
+        return image.copy().astype(np.float32) / 255., self.get_label(idx)
 
     def get_label(self, idx):
         label = self._get_raw_labels()[self._raw_idx[idx]]
